Change timer detection behavior to check every 0.5s
The detection function was being cleared once detected, but that means
the timer won't be added back when the user clicks on the discussions
page and clicks back to problem/solution/submissions page. Now it will
keep checking for the timer eveyr 0.5s and make sure the timer is added
back.